+++ livekit-plugins/livekit-plugins-playai/livekit/plugins/playai/tts.py
@@ -4,10 +4,10 @@ import asyncio
 import os
 import weakref
 from dataclasses import dataclass, fields
+from typing import Optional
 
 from livekit import rtc
 from livekit.agents import (
-    DEFAULT_API_CONNECT_OPTIONS,
     APIConnectionError,
     APIConnectOptions,
     tokenize,
@@ -112,23 +112,22 @@ class TTS(tts.TTS):
             updates["voice"] = voice
         if language is not None:
             updates["language"] = Language(language)
-        tts_kwargs = {k: v for k, v in kwargs.items()}
+        updates.update(kwargs)
 
-        self._config = _update_options(self._config, **updates, **tts_kwargs)
+        _validate_kwargs(updates)
+
+        for key, value in updates.items():
+            if value is not None:
+                setattr(self._config, key, value)
 
         if model is not None:
             self._opts.model = model
-
-        for stream in self._streams:
-            stream._config = _update_options(stream._config, **updates, **tts_kwargs)
-            if model is not None:
-                stream._opts.model = model
 
     def synthesize(
         self,
         text: str,
         *,
-        conn_options: APIConnectOptions = DEFAULT_API_CONNECT_OPTIONS,
+        conn_options: Optional[APIConnectOptions] = None,
     ) -> "ChunkedStream":
         return ChunkedStream(
             tts=self,
@@ -138,7 +137,7 @@ class TTS(tts.TTS):
         )
 
     def stream(
-        self, *, conn_options: APIConnectOptions = DEFAULT_API_CONNECT_OPTIONS
+        self, *, conn_options: Optional[APIConnectOptions] = None
     ) -> "SynthesizeStream":
         stream = SynthesizeStream(
             tts=self,
@@ -155,8 +154,8 @@ class ChunkedStream(tts.ChunkedStream):
         *,
         tts: TTS,
         input_text: str,
-        conn_options: APIConnectOptions,
         opts: _Options,
+        conn_options: Optional[APIConnectOptions] = None,
     ) -> None:
         super().__init__(tts=tts, input_text=input_text, conn_options=conn_options)
         self._client = tts._client
@@ -199,8 +198,8 @@ class SynthesizeStream(tts.SynthesizeStream):
         self,
         *,
         tts: TTS,
-        conn_options: APIConnectOptions,
         opts: _Options,
+        conn_options: Optional[APIConnectOptions] = None,
     ):
         super().__init__(tts=tts, conn_options=conn_options)
         self._client = tts._client
@@ -281,14 +280,6 @@ class SynthesizeStream(tts.SynthesizeStream):
                     yield word.token
 
         return text_stream()
-
-
-def _update_options(config: TTSOptions, **kwargs) -> TTSOptions:
-    _validate_kwargs(kwargs)
-    for k, v in kwargs.items():
-        if v is not None:
-            setattr(config, k, v)
-    return config
 
 
 def _validate_kwargs(kwargs: dict) -> None: